+++ backend/app/controllers/Documentos.java
@@ -16,19 +16,19 @@ import java.io.IOException;
 public class Documentos extends InternalController {
 
    public static void download(Long id) {
+
+		verificarPermissao(Acao.BAIXAR_DOCUMENTO);
+
+		returnIfNull(id, "Long");
+
+		Documento documento = Documento.findById(id);
 	   
-	   verificarPermissao(Acao.VALIDAR_PARECER_GEO,  Acao.INICIAR_PARECER_GEO);
-	   
-	   returnIfNull(id, "Long");
-	   
-	   Documento documento = Documento.findById(id);
-	   
-	   if(documento != null) {
+		if(documento != null) {
 		   File documentoBinary = documento.getFile();
 		   renderBinary(documentoBinary, documentoBinary.getName());
-	   }
+		}
 
-	   renderMensagem(Mensagem.DOCUMENTO_NAO_ENCONTRADO);
+		renderMensagem(Mensagem.DOCUMENTO_NAO_ENCONTRADO);
 	   
    }