+++ frontend/src/utils/tiposDocumentos.js
@@ -8,7 +8,8 @@ var TiposDocumentosAnalise = {
     PARECER_ANALISE_GEO: 16,
     NOTIFICACAO: 22,
     DOCUMENTO_ANALISE_TEMPORAL:23,
-    INCONSISTENCIA_TECNICA:26
+    INCONSISTENCIA_TECNICA:26,
+    AUTO_INFRACAO: 29
 };
 
 exports.utils.TiposDocumentosAnalise = TiposDocumentosAnalise;